requirements.txt
scikit-learn~=1.2.2
pandas~=1.5.3
jupyter~=1.0.0
jupyter_contrib_nbextensions~=0.7.0
apache-airflow~=2.4.3
attrs~=22.1.0
numpy~=1.24.3
pyyaml~=6.0
docker~=6.0.1
mlflow~=2.3.2
matplotlib~=3.7.1
dvc-s3
seaborn~=0.12.2
fsspec
catboost
Pillow~=9.4.0
argparse~=1.4.0